Move NumericArrowsInput to separate component
