+++ src/components/Slider/CounterSlider.tsx
@@ -1,10 +1,6 @@
 import { FC } from 'react'
 
-import classNames from 'classnames'
-
-import { InputArrowUp } from '@banx/icons/InputArrowUp'
-
-import NumericInput from '../inputs/NumericInput'
+import NumericArrowsInput from '../inputs/NumericArrowsInput'
 import { Slider, SliderProps } from './Slider'
 
 import styles from './Slider.module.less'
@@ -17,21 +13,6 @@ export const CounterSlider: FC<SliderProps> = ({
   labelClassName,
   rootClassName,
 }) => {
-  const canIncrement = value < max
-  const canDecrement = value > 0
-
-  const incrementValue = () => {
-    if (canIncrement) {
-      onChange(value + 1)
-    }
-  }
-
-  const decrementValue = () => {
-    if (canDecrement) {
-      onChange(value - 1)
-    }
-  }
-
   return (
     <div className={styles.couterSliderContainer}>
       <Slider
@@ -44,30 +25,7 @@ export const CounterSlider: FC<SliderProps> = ({
         labelClassName={labelClassName}
         rootClassName={rootClassName}
       />
-      <div className={styles.counterInputContainer}>
-        <NumericInput
-          className={styles.counterInput}
-          value={String(value)}
-          onChange={(value) => onChange(Number(value))}
-          positiveOnly
-          integerOnly
-        />
-        <div className={styles.customCounterControls}>
-          <InputArrowUp
-            className={classNames(styles.arrow, { [styles.disabled]: !canIncrement })}
-            onClick={incrementValue}
-          />
-          <div className={styles.separatorLine} />
-          <InputArrowUp
-            onClick={decrementValue}
-            className={classNames(
-              styles.arrow,
-              { [styles.disabled]: !canDecrement },
-              { [styles.rotate]: true },
-            )}
-          />
-        </div>
-      </div>
+      <NumericArrowsInput onChange={onChange} value={value} max={max} />
     </div>
   )
 }
