+++ src/components/Slider/Slider.module.less
@@ -32,59 +32,3 @@
   padding: 0;
   width: 100%;
 }
-
-.counterInputContainer {
-  border: var(--border-primary) !important;
-  max-width: 76px;
-  min-width: 58px;
-  position: relative;
-
-  input {
-    height: 38px;
-
-    @media (max-width: 640px) {
-      padding: 8px 24px 8px 4px !important;
-    }
-  }
-}
-
-.customCounterControls {
-  border-left: var(--border-primary);
-
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  justify-content: space-evenly;
-
-  position: absolute;
-  top: 0;
-  right: 0;
-  height: 100%;
-  width: 22px;
-}
-
-.separatorLine {
-  background: var(--bg-border);
-  height: 1px;
-  width: 100%;
-}
-
-.arrow {
-  cursor: pointer;
-
-  path {
-    fill: var(--content-primary);
-  }
-
-  &.rotate {
-    transform: rotate(180deg);
-  }
-
-  &.disabled {
-    cursor: default;
-
-    path {
-      fill: var(--content-secondary);
-    }
-  }
-}
